Remove update config for Solr
Solr does not behave well with rolling updates due to file lock mechanism. However, if we set up from the update_config property a min 0 and max 100, compose cli ecs integration will not correctly translate it into the generated CF template, and upon deployment of the template you'll get an error stating both min and max percent cannot be 100. Because compose cli will omit the property that has a value of 0. Thus, this needs to be configured manually after deployment.